ProjectController's done

+++ app/Http/Controllers/API/ProjectController.php
@@ -8,6 +8,7 @@ use App\Models\Project;
 use Illuminate\Http\Request;
 use App\Libs\Response;
 use Str;
+use DB;
 
 class ProjectController extends Controller
 {
@@ -23,19 +24,22 @@ class ProjectController extends Controller
      */
     public function index()
     {
+        $projects = Project::with(['owner' => function($owner){
+                                    $owner->with(['profile']);
+                                }, 'tags']);
+
         // if user want to show only the project using technology they used
         if(request()->query('technologies')){
-            $projects = Project::with(['owner', 'tags'])
-                                ->whereHas('tags', function($tag){
+            $projects = $projects->whereHas('tags', function($tag){
                                     $tag->whereIn('tag_id', request()->query('technologies'));
                                 })
                                 ->latest()
                                 ->get();
         } else {
-            $projects = Project::with(['owner', 'tags'])->latest()->get();
+            $projects = $projects->latest()->get();
         }
 
-        Response::successWithData('Project has been loaded', $projects);
+        return Response::successWithData('Project has been loaded', $projects);
     }
 
     /**
@@ -68,6 +72,10 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
+        $project = $project->load(['owner' => function($owner){
+                                $owner->with(['profile']);
+                            }, 'tags']);
+
         return Response::successWithData('Project has been loaded', $project);
     }
 
@@ -93,11 +101,11 @@ class ProjectController extends Controller
             $requests['logo'] = $request->file('logo')->storeOnCloudinaryAs('projects', Str::slug($request->name));
         }
 
-        // delete project tags
-        $project->tags()->delete();
-
         // update the project
         $project->update($requests);
+
+        // delete project tags
+        DB::table('project_tags')->whereProjectId($project->id)->delete();
 
         // add again project tags
         $project->tags()->attach($request->tags);
@@ -113,17 +121,21 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        if($project->owner_id !== auth()->user()->id){
+            return Response::forbidden("Only owner can delete this project");
+        }
+
         // delete project logo
         if($project->logo !== NULL){
             cloudinary()->destroy('projects/' . $project->slug);
             
-            $project->update([
-                'logo' => NULL
-            ]);
+            // $project->update([
+            //     'logo' => NULL
+            // ]);
         }
 
+        DB::table('project_tags')->whereProjectId($project->id);
         $project->delete();
-        $project->tags()->delete();
 
         return Response::success('Project has been deleted');
     }
